requirements.txt
aiogram
tortoise-orm
pydantic
deepdiff
PyYAML
toml
pytz